requirements.txt
click
pyyaml
gradio
pandas
loguru
beautifulsoup4
elasticsearch
langchain-huggingface
langchain-openai
sentence-transformers